+++ src/deprecated.js
@@ -84,9 +84,12 @@ import { XrInputSource } from './xr/xr-input-source.js';
 import { Controller } from './input/controller.js';
 import { ElementInput } from './input/element-input.js';
 import { GamePads } from './input/game-pads.js';
-import { Keyboard, KeyboardEvent } from './input/keyboard.js';
-import { Mouse, MouseEvent } from './input/mouse.js';
-import { getTouchTargetCoords, Touch, TouchDevice, TouchEvent } from './input/touch.js';
+import { Keyboard } from './input/keyboard.js';
+import { KeyboardEvent } from './input/keyboard-event.js';
+import { Mouse } from './input/mouse.js';
+import { MouseEvent } from './input/mouse-event.js';
+import { TouchDevice } from './input/touch-device.js';
+import { getTouchTargetCoords, Touch,TouchEvent } from './input/touch-event.js';
 
 import { FILLMODE_FILL_WINDOW, FILLMODE_KEEP_ASPECT, FILLMODE_NONE, RESOLUTION_AUTO, RESOLUTION_FIXED } from './framework/constants.js';
 import { Application } from './framework/application.js';
